streamlit_app/requirements.txt
streamlit>=1.33
matplotlib
scikit-image
numpy
pandas
Pillow
hydra-core>=1.3
PyYAML
rich
torch
torchvision
